add check column

+++ db/migrate/20131107074108_add_sanitazed_phones_to_user.rb
@@ -2,7 +2,7 @@ class AddSanitazedPhonesToUser < ActiveRecord::Migration
 
   def up
 
-    add_column(:users, :sanitized_phones, :string)
+    add_column(:users, :sanitized_phones, :string) unless column_exists?(:users, :sanitized_phones)
 
     for user in User.all
       user.update_column(:sanitized_phones, (user.phone || '').gsub(/[-()\s]/, ''))
@@ -11,7 +11,7 @@ class AddSanitazedPhonesToUser < ActiveRecord::Migration
   end
 
   def down
-    remove_column(:users, :sanitized_phones)
+    remove_column(:users, :sanitized_phones) if column_exists?(:users, :sanitized_phones)
   end
 
 end
